Fixed circular dependecy
