+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "quill-delta-to-html",
-  "version": "0.12.0",
+  "version": "0.12.1",
   "description": "Converts Quill's delta ops to HTML",
   "main": "./dist/commonjs/main.js",
   "types": "./dist/commonjs/main.d.ts",
